requirements.txt
# Core dependencies for Azure Live Interpreter Agent
livekit-agents>=0.8.0
azure-cognitiveservices-speech>=1.40.0
aiohttp>=3.9.0
python-dotenv>=1.0.0

# Optional: For better performance
uvloop>=0.19.0

# Optional: For enhanced logging
structlog>=23.1.0
